requirements.txt
django==5.0.4
djangorestframework==3.15.1
django-filter==25.1
djangorestframework-simplejwt
django-cors-headers